--- src/components/Friend/index.tsx
@@ -1,3 +1,0 @@
-import Friend from './Friend';
-
-export default Friend;
