+++ src/components/friend/FriendContent/EmptyContent/EmptyContent.css.ts
