+++ src/components/friend/FriendContent/EmptyContent/EmptyContent.tsx
