+++ src/components/friend/FriendContent/EmptyContent/index.ts
