+++ src/components/friend/FriendContent/FriendContent.css.ts
