+++ src/components/friend/FriendContent/index.tsx
@@ -1,0 +1,3 @@
+import FriendContent from './FriendContent';
+
+export default FriendContent;
